+++ client/src/router/index.js
@@ -1,6 +1,7 @@
 import { createRouter, createWebHistory } from "vue-router";
 import CanvasView from "../views/CanvasView.vue";
 import PriceChartView from "../views/PriceChartView.vue";
+import LeaderboardView from "../views/LeaderboardView.vue";
 
 const routes = [
   {
@@ -8,6 +9,7 @@ const routes = [
     component: CanvasView,
   },
   { path: "/pricechart/", component: PriceChartView },
+  { path: "/leaderboard/", component: LeaderboardView },
 ];
 
 const router = createRouter({